+++ gradio/blocks.py
@@ -56,7 +56,13 @@ from gradio.context import (
     get_render_context,
     set_render_context,
 )
-from gradio.data_classes import BlocksConfigDict, FileData, GradioModel, GradioRootModel
+from gradio.data_classes import (
+    BlocksConfigDict,
+    DeveloperPath,
+    FileData,
+    GradioModel,
+    GradioRootModel,
+)
 from gradio.events import (
     EventData,
     EventListener,
@@ -409,7 +415,7 @@ class BlockContext(Block):
             render=render,
         )
 
-    TEMPLATE_DIR = "./templates/"
+    TEMPLATE_DIR = DeveloperPath("./templates/")
     FRONTEND_DIR = "../../frontend/"
 
     @property
@@ -661,16 +667,20 @@ class BlocksConfig:
         self,
         targets: Sequence[EventListenerMethod],
         fn: Callable | None,
-        inputs: Component
-        | BlockContext
-        | Sequence[Component | BlockContext]
-        | AbstractSet[Component | BlockContext]
-        | None,
-        outputs: Component
-        | BlockContext
-        | Sequence[Component | BlockContext]
-        | AbstractSet[Component | BlockContext]
-        | None,
+        inputs: (
+            Component
+            | BlockContext
+            | Sequence[Component | BlockContext]
+            | AbstractSet[Component | BlockContext]
+            | None
+        ),
+        outputs: (
+            Component
+            | BlockContext
+            | Sequence[Component | BlockContext]
+            | AbstractSet[Component | BlockContext]
+            | None
+        ),
         preprocess: bool = True,
         postprocess: bool = True,
         scroll_to_output: bool = False,
@@ -1018,6 +1028,7 @@ class Blocks(BlockContext, BlocksEvents, metaclass=BlocksMeta):
                 t.start()
         else:
             os.environ["HF_HUB_DISABLE_TELEMETRY"] = "True"
+        self.enable_monitoring: bool | None = None
 
         self.default_config = BlocksConfig(self)
         super().__init__(render=False, **kwargs)
@@ -1225,6 +1236,7 @@ class Blocks(BlockContext, BlocksEvents, metaclass=BlocksMeta):
                 dependency.pop("zerogpu", None)
                 dependency.pop("id", None)
                 dependency.pop("rendered_in", None)
+                dependency.pop("every", None)
                 dependency["preprocess"] = False
                 dependency["postprocess"] = False
                 if is_then_event:
@@ -2154,7 +2166,9 @@ Received outputs:
         share: bool | None = None,
         debug: bool = False,
         max_threads: int = 40,
-        auth: Callable | tuple[str, str] | list[tuple[str, str]] | None = None,
+        auth: (
+            Callable[[str, str], bool] | tuple[str, str] | list[tuple[str, str]] | None
+        ) = None,
         auth_message: str | None = None,
         prevent_thread_lock: bool = False,
         show_error: bool = False,
@@ -2180,7 +2194,7 @@ Received outputs:
         auth_dependency: Callable[[fastapi.Request], str | None] | None = None,
         max_file_size: str | int | None = None,
         _frontend: bool = True,
-        enable_monitoring: bool = False,
+        enable_monitoring: bool | None = None,
     ) -> tuple[FastAPI, str, str]:
         """
         Launches a simple web server that serves the demo. Can also be used to create a
@@ -2216,6 +2230,7 @@ Received outputs:
             share_server_protocol: Use this to specify the protocol to use for the share links. Defaults to "https", unless a custom share_server_address is provided, in which case it defaults to "http". If you are using a custom share_server_address and want to use https, you must set this to "https".
             auth_dependency: A function that takes a FastAPI request and returns a string user ID or None. If the function returns None for a specific request, that user is not authorized to access the app (they will see a 401 Unauthorized response). To be used with external authentication systems like OAuth. Cannot be used with `auth`.
             max_file_size: The maximum file size in bytes that can be uploaded. Can be a string of the form "<value><unit>", where value is any positive integer and unit is one of "b", "kb", "mb", "gb", "tb". If None, no limit is set.
+            enable_monitoring: Enables traffic monitoring of the app through the /monitoring endpoint. By default is None, which enables this endpoint. If explicitly True, will also print the monitoring URL to the console. If False, will disable monitoring altogether.
         Returns:
             app: FastAPI app object that is running the demo
             local_url: Locally accessible link to the demo
@@ -2434,6 +2449,7 @@ Received outputs:
             print(
                 f"Monitoring URL: {self.local_url}monitoring/{self.app.analytics_key}"
             )
+        self.enable_monitoring = enable_monitoring in [True, None]
 
         # If running in a colab or not able to access localhost,
         # a shareable link must be created.