+++ js/wasm/src/worker-proxy.ts
@@ -38,7 +38,7 @@ export class WorkerProxy extends EventTarget {
 
 		console.debug("WorkerProxy.constructor(): Create a new worker.");
 		// Loading a worker here relies on Vite's support for WebWorkers (https://vitejs.dev/guide/features.html#web-workers),
-		// assuming that this module is imported from the Gradio frontend (`@gradio/app`), which is bundled with Vite.
+		// assuming that this module is imported from the Gradio frontend (`@gradio/lite`), which is bundled with Vite.
 		// HACK: Use `CrossOriginWorkerMaker` imported as `Worker` here.
 		// Read the comment in `cross-origin-worker.ts` for the detail.
 		const workerMaker = new Worker(new URL("./webworker.js", import.meta.url), {