+++ js/html/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gradio/html
 
+## 0.3.4
+
+### Dependency updates
+
+- @gradio/atoms@0.7.9
+- @gradio/statustracker@0.7.4
+- @gradio/icons@0.7.0
+
 ## 0.3.3
 
 ### Dependency updates