+++ js/number/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/number
 
+## 0.4.11
+
+### Dependency updates
+
+- @gradio/atoms@0.7.9
+- @gradio/statustracker@0.7.4
+
 ## 0.4.10
 
 ### Dependency updates